requirements.txt
# Core Framework
Flask>=3.0.0
Werkzeug>=3.0.0

# Machine Learning & Deep Learning
tensorflow>=2.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-learn>=1.3.0

# Data Processing
pandas>=2.1.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Image Processing & Augmentation
albumentations>=1.3.0
imgaug>=0.4.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
requests>=2.31.0

# Development & Testing (Optional)
# jupyter>=1.0.0
# notebook>=7.0.0

# Production deployment (Optional)
# gunicorn>=21.2.0
# eventlet>=0.33.3